Upgraded React Native to v0.61.4, Updated dependencies
