+++ code/mobile/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.mobile",
+    package = "com.crate",
 )
 
 android_resource(
     name = "res",
-    package = "com.mobile",
+    package = "com.crate",
     res = "src/main/res",
 )
