+++ code/mobile/android/app/src/main/java/com/crate/MainActivity.java
@@ -1,0 +1,15 @@
+package com.crate;
+
+import com.facebook.react.ReactActivity;
+
+public class MainActivity extends ReactActivity {
+
+  /**
+   * Returns the name of the main component registered from JavaScript. This is used to schedule
+   * rendering of the component.
+   */
+  @Override
+  protected String getMainComponentName() {
+    return "crate";
+  }
+}
