+++ code/mobile/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "mobile",
+  "name": "crate",
   "displayName": "crate"
 }
