+++ code/mobile/ios/crate.xcodeproj/xcshareddata/xcschemes/crate-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "mobile-tvOS.app"
-               BlueprintName = "mobile-tvOS"
-               ReferencedContainer = "container:mobile.xcodeproj">
+               BuildableName = "crate-tvOS.app"
+               BlueprintName = "crate-tvOS"
+               ReferencedContainer = "container:crate.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "mobile-tvOSTests.xctest"
-               BlueprintName = "mobile-tvOSTests"
-               ReferencedContainer = "container:mobile.xcodeproj">
+               BuildableName = "crate-tvOSTests.xctest"
+               BlueprintName = "crate-tvOSTests"
+               ReferencedContainer = "container:crate.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "mobile-tvOSTests.xctest"
-               BlueprintName = "mobile-tvOSTests"
-               ReferencedContainer = "container:mobile.xcodeproj">
+               BuildableName = "crate-tvOSTests.xctest"
+               BlueprintName = "crate-tvOSTests"
+               ReferencedContainer = "container:crate.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "mobile-tvOS.app"
-            BlueprintName = "mobile-tvOS"
-            ReferencedContainer = "container:mobile.xcodeproj">
+            BuildableName = "crate-tvOS.app"
+            BlueprintName = "crate-tvOS"
+            ReferencedContainer = "container:crate.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "mobile-tvOS.app"
-            BlueprintName = "mobile-tvOS"
-            ReferencedContainer = "container:mobile.xcodeproj">
+            BuildableName = "crate-tvOS.app"
+            BlueprintName = "crate-tvOS"
+            ReferencedContainer = "container:crate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "mobile-tvOS.app"
-            BlueprintName = "mobile-tvOS"
-            ReferencedContainer = "container:mobile.xcodeproj">
+            BuildableName = "crate-tvOS.app"
+            BlueprintName = "crate-tvOS"
+            ReferencedContainer = "container:crate.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
