+++ code/mobile/ios/crate/AppDelegate.h
