+++ code/mobile/ios/crate/Images.xcassets/AppIcon.appiconset/Contents.json
