+++ code/mobile/ios/crate/Images.xcassets/Contents.json
