+++ code/mobile/ios/crate/Info.plist
@@ -1,0 +1,63 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CFBundleDevelopmentRegion</key>
+	<string>en</string>
+	<key>CFBundleDisplayName</key>
+	<string>crate</string>
+	<key>CFBundleExecutable</key>
+	<string>$(EXECUTABLE_NAME)</string>
+	<key>CFBundleIdentifier</key>
+	<string>$(PRODUCT_BUNDLE_IDENTIFIER)</string>
+	<key>CFBundleInfoDictionaryVersion</key>
+	<string>6.0</string>
+	<key>CFBundleName</key>
+	<string>$(PRODUCT_NAME)</string>
+	<key>CFBundlePackageType</key>
+	<string>APPL</string>
+	<key>CFBundleShortVersionString</key>
+	<string>1.0</string>
+	<key>CFBundleSignature</key>
+	<string>????</string>
+	<key>CFBundleVersion</key>
+	<string>1</string>
+	<key>LSRequiresIPhoneOS</key>
+	<true/>
+	<key>NSAppTransportSecurity</key>
+  	<dict>
+  		<key>NSAllowsArbitraryLoads</key>
+  		<true/>
+  		<key>NSExceptionDomains</key>
+  		<dict>
+  			<key>localhost</key>
+  			<dict>
+  				<key>NSExceptionAllowsInsecureHTTPLoads</key>
+  				<true/>
+  			</dict>
+  		</dict>
+  	</dict>
+  	<key>NSLocationWhenInUseUsageDescription</key>
+  	<string></string>
+  	<key>UIAppFonts</key>
+  	<array>
+  		<string>MaterialCommunityIcons.ttf</string>
+  		<string>MaterialIcons.ttf</string>
+  		<string>Zocial.ttf</string>
+  	</array>
+  	<key>UILaunchStoryboardName</key>
+  	<string>LaunchScreen</string>
+  	<key>UIRequiredDeviceCapabilities</key>
+  	<array>
+  		<string>armv7</string>
+  	</array>
+  	<key>UISupportedInterfaceOrientations</key>
+  	<array>
+  		<string>UIInterfaceOrientationPortrait</string>
+  		<string>UIInterfaceOrientationLandscapeLeft</string>
+  		<string>UIInterfaceOrientationLandscapeRight</string>
+  	</array>
+  	<key>UIViewControllerBasedStatusBarAppearance</key>
+  	<false/>
+</dict>
+</plist>
