+++ code/mobile/ios/crateTests/Info.plist
