+++ code/mobile/package.json
@@ -3,39 +3,40 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
+    "android": "react-native run-android",
+    "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint .",
-    "postinstall": "npx jetify"
+    "lint": "eslint ."
   },
   "dependencies": {
-    "@react-native-community/async-storage": "^1.6.1",
+    "@react-native-community/async-storage": "^1.6.3",
     "axios": "^0.19.0",
     "gql-query-builder": "^3.1.3",
     "prop-types": "^15.7.2",
-    "react": "16.8.6",
-    "react-native": "0.60.5",
-    "react-native-gesture-handler": "^1.4.1",
+    "react": "16.9.0",
+    "react-native": "0.61.5",
+    "react-native-gesture-handler": "^1.5.1",
     "react-native-keyboard-aware-scroll-view": "^0.9.1",
-    "react-native-reanimated": "^1.2.0",
+    "react-native-reanimated": "^1.4.0",
     "react-native-vector-icons": "^6.6.0",
-    "react-navigation": "^4.0.6",
-    "react-navigation-stack": "^1.8.0",
-    "react-navigation-tabs": "^2.5.4",
-    "react-redux": "^7.1.1",
+    "react-navigation": "^4.0.10",
+    "react-navigation-stack": "^1.10.3",
+    "react-navigation-tabs": "^2.6.0",
+    "react-redux": "^7.1.3",
     "redux": "^4.0.4",
     "redux-thunk": "^2.3.0",
-    "validator": "^11.1.0"
+    "validator": "^12.1.0"
   },
   "devDependencies": {
-    "@babel/core": "7.6.0",
-    "@babel/runtime": "7.6.0",
-    "@react-native-community/eslint-config": "0.0.3",
+    "@babel/core": "7.7.4",
+    "@babel/runtime": "7.7.4",
+    "@react-native-community/eslint-config": "0.0.5",
     "babel-jest": "24.9.0",
-    "eslint": "6.4.0",
+    "eslint": "6.7.0",
     "jest": "24.9.0",
-    "metro-react-native-babel-preset": "0.54.1",
-    "react-test-renderer": "16.8.6"
+    "metro-react-native-babel-preset": "0.56.3",
+    "react-test-renderer": "16.9.0"
   },
   "jest": {
     "preset": "react-native"
